examples/flask/gdrive-llm/requirements.txt
flask==3.1.0
google-auth-oauthlib==1.2.1
google-auth-httplib2==0.2.0
google-api-python-client==2.157.0
sqlalchemy==2.0.36
openai==1.59.5
celery==5.4.0
python-dotenv==1.0.1
sqlite-vec==0.1.6
gunicorn==23.0.0